// src/utils/Models/Cast.tsx
export type Cast = {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    cast_id: number;
    character: string;
    credit_id: string;
    order: number;
  }
  
  export type Crew = {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    credit_id: string;
    department: string;
    job: string;
  }
  
  export type CrewCastDetail ={
    id: number;
    cast: Cast[];
    crew: Crew[];
  }
  
  
  export type CastDetail ={
    title: String;
    data: String[]
  }
  